image-analysis/requirements.txt
numpy==2.1.1
opencv_python==4.9.0.80
Pillow==9.5.0
Pillow==10.4.0
pyiqa==0.1.12
selenium==4.24.0
sentence_transformers==2.2.2
skimage==0.0
torch==2.1.2
